+++ Felix/ScintillaEdit_test.cpp
@@ -37,56 +37,56 @@ BOOST_AUTO_TEST_SUITE( TestCScintillaEdit )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.html")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:html") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_xml )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.xml")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:xml") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_php )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.php")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:php") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_py )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.py")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:python") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_css )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.css")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:css") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_cpp )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.cpp")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:cpp") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_perl )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.pl")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:perl") ;
 	}
 	BOOST_AUTO_TEST_CASE( set_lexer_js )
 	{
 		CScintillaEdit edit ;
 		edit.set_correct_lexer(_T("foo.js")) ;
-		BOOST_CHECK_EQUAL(1, (int)edit.m_sensing_variable.size()) ;
+		BOOST_CHECK_EQUAL(1u, edit.m_sensing_variable.size()) ;
 		BOOST_CHECK_EQUAL(string(edit.m_sensing_variable[0].c_str()), "lexer:js") ;
 	}
